requirements.txt
# Django Core
django
djangorestframework
psycopg2-binary
python-dotenv
gunicorn

# JWT Authentication
djangorestframework-simplejwt

# PDF Generation (for later)
weasyprint

# AI Integration
openai
google-generativeai
google-genai

# Testing
pytest
pytest-django
